requirements.txt
pytest
pytest-playwright
playwright
PyYAML
Faker
python-dateutil
